# src/docs/swagger.yaml
definitions:
...
  dto.CreatePersianYearRequest:
    properties:
      endAt:
        type: string
      persianTitle:
        maxLength: 4
        minLength: 4
        type: string
      startAt:
        type: string
      year:
        type: integer
    type: object
...
  dto.UpdatePersianYearRequest:
    properties:
      endAt:
        type: string
      persianTitle:
        maxLength: 4
        minLength: 4
        type: string
      startAt:
        type: string
      year:
        type: integer
    type: object
...
  filter.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        description: contains notContains equals notEqual startsWith lessThan lessThanOrEqual
          greaterThan greaterThanOrEqual inRange endsWith
        type: string
    type: object
  filter.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/filter.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/filter.Sort'
        type: array
    type: object
  filter.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
...
  helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/validation.ValidationError'
        type: array
    type: object
  helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  validation.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/years/:
    post:
      consumes:
      - application/json
      description: Create a PersianYear
      parameters:
      - description: Create a PersianYear
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePersianYearRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a PersianYear
      tags:
      - PersianYears
  /v1/years/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a PersianYear
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a PersianYear
      tags:
      - PersianYears
    get:
      consumes:
      - application/json
      description: Get a PersianYear
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a PersianYear
      tags:
      - PersianYears
    put:
      consumes:
      - application/json
      description: Update a PersianYear
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a PersianYear
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePersianYearRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a PersianYear
      tags:
      - PersianYears
  /v1/years/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get PersianYears
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/filter.PaginationInputWithFilter'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get PersianYears
      tags:
      - PersianYears
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"